refactor(decopilot): inline the hosted harness
