+++ apps/api/src/core/studio-context.ts
@@ -16,7 +16,6 @@ import type { Kysely } from "kysely";
 import type { CredentialVault } from "../encryption/credential-vault";
 import type { Database, Permission } from "../storage/types";
 import type { AccessControl } from "./access-control";
-import type { HarnessContext } from "./harness-context";
 export type { BetterAuthInstance } from "@/auth";
 // Re-export for consumers
 export type { AccessControl, CredentialVault };
@@ -383,7 +382,7 @@ export interface Timings {
  * This provides access to all necessary services without coupling
  * to implementation details.
  */
-export interface StudioContext extends HarnessContext {
+export interface StudioContext {
   // Connection ID (from url)
   connectionId?: string;
 
@@ -463,13 +462,9 @@ export interface StudioContext extends HarnessContext {
   ) => Promise<FireAutomationOutcome>;
 
   /**
-   * Sandbox dispatch preference for the in-flight run, populated by
-   * `prepareRun` from the resolved `DispatchTarget`:
-   *   - `"agent-sandbox"` — force hosted sandbox provider behavior for this run.
-   *   - `"cluster-default"` — legacy/env-default hint: use whichever sandbox
-   *     kind `STUDIO_SANDBOX_PROVIDER` resolves to.
-   *   - `"user-desktop"` — use the user's link daemon for sandbox provider
-   *     behavior.
+   * Sandbox dispatch preference for the in-flight run. This remains on the
+   * context until every lifecycle caller has moved to the fixed hosted
+   * provider later in the stack.
    */
   sandboxPreference?: "agent-sandbox" | "cluster-default" | "user-desktop";
 }
